# doc/api.yaml
openapi: 3.0.0
info:
  title: 'WASAText API'
  description: 'API for managing a messaging platform'
  version: 1.0.0
servers:                                              # Base URL where the API is
  - url: 'http://localhost:3000'
    description: 'Local Development Environment (Docker/Go)'
  - url: 'https://api.wasatext.com'
    description: 'Production Server'
security:                                             # Global security
  - BearerAuth: []
components:                                           # Reusable components
  schemas:                                            # Structures
    Error:
      description: 'Standard error'
      type: object
      properties:
        code:
          description: 'Integer code error'
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          readOnly: true
        message:
          description: 'Error text message'
          type: string
          pattern: '^.{0,500}$'                       # 0-500 chars
          minLength: 0
          maxLength: 500
          readOnly: true
      required: [code, message]
      additionalProperties: false                     # No other properties allowed
      example:
        code: 404
        message: 'The requested resource could not be found'
    Emoji:
      description: |
        Exactly one grapheme cluster, capped at 16 Unicode code points.
        Validation is structural rather than an emoji-catalogue lookup, so any one symbol is accepted.
      type: string
      pattern: '^.{1,16}$'                            # 1 symbol, at most 16 code points
      minLength: 1
      maxLength: 16
      example: '👍'
    Id:
      description: 'Standard unique identifier, only the canonical lowercase form is valid'
      type: string
      format: uuid                                    # Universally Unique Identifier
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: 'c56a4180-65aa-42ec-a945-5fd21dec0538'
...
    UserId:
      $ref: '#/components/schemas/Id'
    ChatId:
      $ref: '#/components/schemas/Id'
    MessageId:
      $ref: '#/components/schemas/Id'
    CommentId:
      $ref: '#/components/schemas/Id'
...
    PhotoURL:
      description: |
        Photo URL, relative to the server: the client already knows the host it is talking to.
        The last segment is the id of the photo: a UUID saved with no extension,
        the content type is detected when the photo is served
      type: string                                    # Relative reference, not a whole URL
      pattern: '^/photos/[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 44
      maxLength: 44
      readOnly: true
      example: '/photos/c56a4180-65aa-42ec-a945-5fd21dec0538'
...
    MessageBase:
      description: 'Base structure for messages and snippets'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/MessageId'
        date:
          description: 'Message creation date in UTC, with millisecond precision'
          type: string
          format: date-time
          example: '2025-11-04T10:32:00.000Z'
        user:
          $ref: '#/components/schemas/UserId'
        state:
          description: |
            Message state, computed and never stored:
            read: no member of the chat is left behind the message,
            received: at least one of them still is.
            A member falls behind a message when the message arrives, and catches up by opening the chat
            or by writing in it, since writing means having seen what is above.
            The sender needs no exception: sending catches it up to its own message.
            A member joins caught up, so it answers for nothing sent before it and a message
            already read never turns back; a member leaving takes its answer with it,
            so a message the leaver alone was holding becomes read.
            The value is a fact about the members and not about the caller,
            so it is the same for everybody looking at the chat
          type: string
          pattern: '^.{4,8}$'                         # 4-8 chars as 'read' or 'received'
          minLength: 4
          maxLength: 8
          enum: [received, read]
          example: 'read'
          # EXTRA: sent: message sent by the user but not received yet by all chat members
      required: [id, date, user, state]
    MessageText:
      description: 'Standard message text content'
      type: string
      pattern: '^.{1,5000}$'                          # 1-5000 chars
      minLength: 1
      maxLength: 5000
      example: 'Hello world!'
    Message:
      description: 'Standard message'
      allOf:                                          # Combines MessageBase with the modifications below
        - $ref: '#/components/schemas/MessageBase'    # Inherit everything from MessageBase
        - type: object                                # Override type to object
          properties:
            content:
              description: 'Content of the message, text and/or photo'
              type: object
              properties:
                text:
                  $ref: '#/components/schemas/MessageText'
                photo:
                  $ref: '#/components/schemas/PhotoURL'
              minProperties: 1                        # At least one of text/photo is present
            comments:
              description: 'The reactions on this message, empty list if none'
              type: array
              items:
                $ref: '#/components/schemas/Comment'
              minItems: 0
              maxItems: 1000
              uniqueItems: true                       # One comment per user, and one id per comment
          required: [content, comments]               # + id, date, user, state already required in MessageBase
...
    Comment:
      description: 'Standard comment to a message'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/CommentId'
        user:
          $ref: '#/components/schemas/UserId'
        emoji:
          $ref: '#/components/schemas/Emoji'
      required: [id, user, emoji]
      additionalProperties: false                     # No other properties allowed
  securitySchemes:                                    # Authentication
    BearerAuth:
      description: 'Authentication token (JWT)'       # JSON Web Token
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ParamChatId:
      name: chatId
      description: 'Chat unique identifier'
      in: path
      schema:
        $ref: '#/components/schemas/ChatId'
      required: true
...
    ParamMessageId:
      name: messageId
      description: 'Message unique identifier'
      in: path
      schema:
        $ref: '#/components/schemas/MessageId'
      required: true

    # EXTRA: add cursor based pagination
  responses:
...
    NoContent:
      description: '204 No Content'
    
    # Client errors
    BadRequest:
      description: '400 Bad Request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: 'The data provided is not valid. Please check constraints and required fields'
    Unauthorized:
      description: '401 Unauthorized'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            message: 'Authentication required. Please include a valid access token'
    Forbidden:
      description: '403 Forbidden'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 403
            message: 'You do not have permission to access this resource'
    NotFound:
      description: '404 Not Found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: 'The requested resource was not found'

    #MethodNotAllowed:
    #  description: '405 Method Not Allowed'
    #  content:
    #    application/json:
    #      schema:
    #        $ref: '#/components/schemas/Error'
    #      example:
    #        code: 405
    #        message: 'The HTTP method used is not supported for this resource'

    # Server errors
    InternalServerError:
      description: '500 Internal Server Error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 500
            message: 'An internal server error occurred'
...
paths:                                                # Rest endpoints
...
  /chats/{chatId}/messages/{messageId}/comments/me: # current user unique comment
    parameters:
      - $ref: '#/components/parameters/ParamChatId'
      - $ref: '#/components/parameters/ParamMessageId'

    put:
      operationId: 'commentMessage'
      description: |
        Sets the authenticated user's reaction on a message in a chat they belong to.
        The message must belong to the chat named in the URL.
        Each member has at most one reaction on a message; another call updates that row's emoji while preserving its comment id.
        Commenting catches the caller up with the chat since writing in it means having seen what is above.
        The returned Message contains its unchanged content plus the complete updated comments list and its state after that catch-up. 
      summary: 'Comment a message'
      tags: ['PrivateChat', 'Group']
      requestBody:
        description: 'Payload containing the new single emoji to set as comment'
        content:
          application/json:
            schema:
              description: 'Object containing the emoji of the comment'
              type: object                            # not a Comment: id and user are readOnly,
              additionalProperties: false             # No other properties allowed by the contract
              properties:                             # the server knows both and the client cannot send them
                emoji:
                  $ref: '#/components/schemas/Emoji'
              required: [emoji]
        required: true
      responses:
        '200':
          description: 'Existing comment updated successfully; returns the message with its updated comments'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: '3fa85f64-5717-4562-b3fc-2c963f66afa6'
                content:
                  text: 'Hello, how are you?'
                date: '2025-11-15T18:45:00.000Z'
                user: 'cf6a4180-65aa-42ec-a945-5fd21dec0538'
                state: 'received'
                comments:
                  - id: 'c52a4180-65aa-42ec-a945-5fd21dec0538'
                    emoji: '👍'
                    user: '1a6a4180-65aa-42ec-a945-5fd21dec0538'
        '201':
          description: 'New comment created successfully; returns the message with its updated comments'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: '3fa85f64-5717-4562-b3fc-2c963f66afa6'
                content:
                  text: 'Hello, how are you?'
                date: '2025-11-15T18:45:00.000Z'
                user: 'cf6a4180-65aa-42ec-a945-5fd21dec0538'
                state: 'received'
                comments:
                  - id: 'c52a4180-65aa-42ec-a945-5fd21dec0538'
                    emoji: '👍'
                    user: '1a6a4180-65aa-42ec-a945-5fd21dec0538'
        '400':                                        # Invalid URL id, body or emoji
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':                                        # The message belongs to this chat, but the caller does not
          $ref: '#/components/responses/Forbidden'
        '404':                                        # No message owns the id, or it belongs to another chat
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      operationId: 'uncommentMessage'
      description: |
        Removes the authenticated user's own reaction from a message in a chat they belong to.
        The message must belong to the chat named in the URL, and the caller must already have a comment on the message in the URL.
        Other members' comments and the message itself are left untouched.
        A successful removal catches the caller up with the chat without moving lastReadDate backwards.
      summary: 'Remove my comment from a message'
      tags: ['PrivateChat', 'Group']
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':                                        # Invalid chatId or messageId
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':                                        # The message belongs to this chat, but the caller does not
          $ref: '#/components/responses/Forbidden'
        '404':                                        # Message/chat mismatch, missing message, or no caller comment
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...